backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==5.0.0
black==25.1.0
boto3==1.39.4
botocore==1.39.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
#emergent-plugins @ file:///wheelhouse/emergent_plugins-1.0.0-py3-none-any.whl#sha256=708b99f3e454ea78bd5d97c01cc1754b57e5f758bd46a23dc194403c7e77bab4
fastapi==0.110.1
flake8==7.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jmespath==1.0.1
jq==1.9.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mypy==1.16.1
mypy_extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
playwright==1.56.0
pluggy==1.6.0
psutil==7.0.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.37.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.0
fastapi-mail==1.4.1
firebase-admin==6.5.0
